+++ nodes/decoder.js
@@ -614,15 +614,15 @@ module.exports = {
   },
 
   /**
-    * Decodes the response to a set-super-passwords request (function code 0x8c).
+    * Decodes the response to a set-door-passcodes request (function code 0x8c).
     *
     * @param {buffer}   buffer     64 byte NodeJS buffer
     * @param {function} translator (optional) function to internationalise the text in a
     *                              decoded object
     *
-    * @param {object}   Decoded set-super-passwords response object
+    * @param {object}   Decoded set-door-passcodes response object
     */
-  SetSuperPasswords: function (bytes, translator) {
+  SetDoorPasscodes: function (bytes, translator) {
     return {
       deviceId: uint32(bytes, 4),
       ok: bool(bytes, 8)